move fragment transition to a class and clean the MainActivity
